Updated Hostel Management System/requirements.txt
pillow
qrcode[pil]
reportlab
fpdf2